Add user creation functionality to setup_ETS2_server.sh
Implemented a new feature to create a dedicated user for the ETS2 server when the script is run as root. The script now prompts for a username and offers options to generate a random password or set it manually. It includes user management tasks such as adding the user to the sudo group and copying the setup script to the new user's home directory.